filter now updates graph object
since each output can have only one callback, the callback which updates
the Plotly figure is necessarily involved. the inputs and outputs having
to be defined in the callback decorator, only the functions which act on
the data are defined in the external modules for updating the Plotly
figure. currently the filter only supports one value per filter for
discrete values.

global variable changes will generally break Dash apps. so the boolean
filter is reevaluated everytime the figure is updated. global changes
to make "dummy columns" for the data frame have been changed to
initialization in data. other global variable can be preinitialized for
all possible values, or caching optimization can be done. but it is
unlikely boolean filters are a significant cost.

similarly, column aliasing shouldn't change the dataframe object, but
only change the columns displayed and available to select, and the
options also shouldn't be global. changes coming in next commit