removed global variable changes
as in commit c1bd492aa588664b7f76caba96d3ca0f4c76238d, this is to remove
all global variable changes which can break dash apps. however the
aliases used are substituted with the name and only used for lookup,
when it is desired the alias is used for both lookup and display.